Refactor Onboarding components.
- Split OnboardingChecklist into smaller components.
- Moved OnboardingBox to a separate component.
