+++ src/overview/onboarding/components/checklist.css
