+++ src/overview/onboarding/components/index.ts
@@ -1,0 +1,5 @@
+import OnboardingBox from './onboarding-box'
+
+export { default as OnboardingChecklist } from './checklist-container'
+
+export default OnboardingBox
