+++ src/overview/onboarding/components/onboarding-box.css
