+++ src/overview/onboarding/index.js
@@ -1,9 +1,9 @@
 import * as selectors from './selectors'
 import * as actions from './actions'
 import * as constants from './constants'
-import container from './components'
+import OnboardingBox from './components'
 
 export { default as reducer } from './reducer'
 export { selectors, actions, constants }
 
-export default container
+export default OnboardingBox
